+++ src/main/java/clubhouse/clubhouse/domain/application/dto/form/UserInfoForm.java
@@ -1,4 +1,4 @@
-package clubhouse.clubhouse.domain.application.dto;
+package clubhouse.clubhouse.domain.application.dto.form;
 
 import clubhouse.clubhouse.domain.member.entity.Gender;
 import lombok.Data;
