+++ src/main/java/clubhouse/clubhouse/domain/application/dto/response/ApplicationDetailResponseDto.java
@@ -1,4 +1,4 @@
-package clubhouse.clubhouse.domain.application.dto;
+package clubhouse.clubhouse.domain.application.dto.response;
 
 import lombok.Data;
 import org.springframework.http.HttpStatus;
