+++ src/main/java/clubhouse/clubhouse/domain/application/dto/response/ApplyListResponseDto.java
@@ -1,5 +1,6 @@
-package clubhouse.clubhouse.domain.application.dto;
+package clubhouse.clubhouse.domain.application.dto.response;
 
+import clubhouse.clubhouse.domain.application.dto.form.ApplyListResponseForm;
 import lombok.Data;
 import org.springframework.http.HttpStatus;
