+++ src/main/java/clubhouse/clubhouse/domain/application/dto/response/MyPageResponseDto.java
@@ -1,5 +1,6 @@
-package clubhouse.clubhouse.domain.application.dto;
+package clubhouse.clubhouse.domain.application.dto.response;
 
+import clubhouse.clubhouse.domain.application.dto.form.UserApplyListForm;
 import lombok.Data;
 
 import java.util.List;
